BE/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.3.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.3
fastapi-users